\name{chemometrics}
\alias{chemometrics}
\docType{data}
\title{
Chemometrics data set from Sardy (2008)
}
\description{
Fuel octane level measurements with sample size N = 434 and P = 351 spectrometer measurements.
}
\usage{data(chemometrics)}
\format{
  A data frame with 434 observations on the following 2 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x}}{a matrix with 351 columns}
  }
}

\references{
S. Sardy. On the practice of rescaling covariates. International Statistical Review. 2008

}
\examples{
data(chemometrics)
}