\name{internetAd}
\alias{internetAd}
\docType{data}
\title{
InternetAd data set from Kushmerick (1999)
}
\description{
Classification of N = 2359 possible advertisements on internet pages based on P = 1430 features
}
\usage{data(internetAd)}
\format{
  A data frame with 2359 observations on the following 2 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x}}{a matrix with 1430 columns}
  }
}

\references{
N. Kushmerick. Learning to remove internet advertisements. In Proceedings of the Third Annual Conference on Autonomous Agents. 1999

}
\examples{
data(internetAd)
}