\name{leukemia}
\alias{leukemia}
\docType{data}
\title{
Gene expression data from Golub et al. (1999)
}
\description{
Gene expression data (3051 genes and 38 tumor mRNA samples) from the leukemia microarray
study of Golub et al. (1999).
}
\usage{data(leukemia)}
\format{

   A list with 3 variables
  \describe{
    \item{\code{Y}}{a numeric vector with 38 observations}
    \item{\code{X}}{a matrix with 3051 columns}
    \item{\code{gene.names}}{Gene names}
  }
}

\references{
Golub et al. (1999). Molecular classification of cancer: class discovery and class prediction by gene expression monitorin, Science 286, 531-537
}
\examples{
data(leukemia)

}

