\name{riboflavin}
\alias{riboflavin}
\docType{data}
\title{
Riboflavin data set from Buhlmann et al. (2013) 
}
\description{
Dataset of riboflavin production by Bacillus subtilis containing n = 71 observations of p = 4088
predictors (gene expressions) and a one-dimensional response (riboflavin production)
}
\usage{data(riboflavin)}
\format{
  A data frame with 71 observations on the following 2 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x}}{a matrix with 4088 columns}
  }
}

\references{
Buhlmann, P., Kalisch, M. and Meier, L. (2013). High-dimensional statistics with a view towards
applications in biology. To appear in Annual Review of Statistics and its Applications.

}
\examples{
data(riboflavin)
}